+++ packages/kitchensink/service-fetchq-task/feature-home.js
@@ -1,7 +1,0 @@
-const homeRouteHandler = async () => 'fetchq';
-
-module.exports = {
-  name: 'home',
-  target: '$FASTIFY_GET',
-  handler: [{ url: '/', handler: homeRouteHandler }],
-};